Put back lbrynet servers into the config
