+++ lbrytv.yml
@@ -1,3 +1,8 @@
+LbrynetServers:
+  default: http://localhost:5581/
+  lbrynet1: http://localhost:5581/
+  lbrynet2: http://localhost:5581/
+
 Debug: 1
 
 InternalAPIHost: https://api.lbry.com
